apps/requirements.txt
django==4.2.11
djangorestframework==3.14.0
django-filter==23.5
psycopg==3.1.18
psycopg-binary==3.1.18
uWSGI==2.0.24
requests==2.31.0
jsonschema==4.21.0
tzdata==2024.1
django-cors-headers==4.3.0
